Included ROS 2 lint testing

+++ cloudini_ros/test/rosbag_benchmark.cpp
@@ -1,25 +1,56 @@
+// Copyright (c) 2025 Davide Faconti
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//    * Redistributions of source code must retain the above copyright
+//      notice, this list of conditions and the following disclaimer.
+//
+//    * Redistributions in binary form must reproduce the above copyright
+//      notice, this list of conditions and the following disclaimer in the
+//      documentation and/or other materials provided with the distribution.
+//
+//    * Neither the name of the Willow Garage nor the names of its
+//      contributors may be used to endorse or promote products derived from
+//      this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
 #include <stdio.h>
 
 #include <chrono>
 #include <functional>
 #include <iostream>
 #include <memory>
+#include <string>
+
 #include <rclcpp/rclcpp.hpp>
 #include <rclcpp/serialization.hpp>
 #include <rclcpp/serialized_message.hpp>
 #include <rosbag2_cpp/reader.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
-#include <string>
 
 #include "cloudini_lib/cloudini.hpp"
 #include "cloudini_ros/conversion_utils.hpp"
 
-struct StatsData {
-  long total_time_usec = 0;
+struct StatsData
+{
+  int32_t total_time_usec = 0;
   double total_ratio = 0;
 };
 
-struct Statistics {
+struct Statistics
+{
   int count = 0;
   StatsData lz4_only;
   StatsData zstd_only;
@@ -27,7 +58,10 @@ struct Statistics {
   StatsData lossy_zstd;
 };
 
-void compress(const sensor_msgs::msg::PointCloud2& msg, const Cloudini::EncodingInfo& encoding_info, StatsData& stats) {
+void compress(
+  const sensor_msgs::msg::PointCloud2 & msg, const Cloudini::EncodingInfo & encoding_info,
+  StatsData & stats)
+{
   Cloudini::PointcloudEncoder encoder(encoding_info);
   std::vector<uint8_t> compressed_data;
   Cloudini::ConstBufferView input(msg.data.data(), msg.data.size());
@@ -37,10 +71,12 @@ void compress(const sensor_msgs::msg::PointCloud2& msg, const Cloudini::Encoding
   auto t2 = std::chrono::high_resolution_clock::now();
 
   stats.total_time_usec += std::chrono::duration_cast<std::chrono::microseconds>(t2 - t1).count();
-  stats.total_ratio += double(compressed_data.size()) / double(msg.data.size());
+  stats.total_ratio += static_cast<double>(compressed_data.size()) /
+    static_cast<double>(msg.data.size());
 }
 
-int main(int argc, char** argv) {
+int main(int argc, char ** argv)
+{
   if (argc != 2) {
     std::cerr << "Usage: " << argv[0] << " <bag>" << std::endl;
     return 1;
@@ -55,7 +91,7 @@ int main(int argc, char** argv) {
 
   std::unordered_map<std::string, Statistics> statistics_by_topic;
 
-  for (const auto& it : reader.get_all_topics_and_types()) {
+  for (const auto & it : reader.get_all_topics_and_types()) {
     if (it.type == "sensor_msgs/msg/PointCloud2") {
       std::cout << "Found PointCloud2 topic: " << it.name << std::endl;
       statistics_by_topic[it.name] = Statistics();
@@ -73,13 +109,13 @@ int main(int argc, char** argv) {
     auto ros_msg = std::make_shared<sensor_msgs::msg::PointCloud2>();
     serialization.deserialize_message(&serialized_msg, ros_msg.get());
 
-    auto& statistics = statistics_by_topic[msg->topic_name];
+    auto & statistics = statistics_by_topic[msg->topic_name];
     // 100 microns resolution!
     Cloudini::EncodingInfo encoding_info = Cloudini::ConvertToEncodingInfo(*ros_msg, 0.0001F);
 
     if (statistics.count == 0) {
       std::cout << "Topic [" << msg->topic_name << "] has fields:\n";
-      for (const auto& field : encoding_info.fields) {
+      for (const auto & field : encoding_info.fields) {
         std::cout << " - " << field.name << " (" << Cloudini::ToString(field.type) << ")\n";
       }
       std::cout << std::endl;
@@ -105,22 +141,22 @@ int main(int argc, char** argv) {
   }
 
   //------------------------------------------------------------
-  for (const auto& [topic, stat] : statistics_by_topic) {
-    double dcount = double(stat.count);
+  for (const auto & [topic, stat] : statistics_by_topic) {
+    double dcount = static_cast<double>(stat.count);
     std::cout << "Topic: " << topic << std::endl;
     std::cout << "  Count: " << stat.count << std::endl;
     printf(
-        "  [LZ4 only]      ratio: %.2f time (usec): %ld\n", stat.lz4_only.total_ratio / dcount,
+        "  [LZ4 only]      ratio: %.2f time (usec): %d\n", stat.lz4_only.total_ratio / dcount,
         stat.lz4_only.total_time_usec / stat.count);
     printf(
-        "  [ZSTD only]     ratio: %.2f time (usec): %ld\n", stat.zstd_only.total_ratio / dcount,
+        "  [ZSTD only]     ratio: %.2f time (usec): %d\n", stat.zstd_only.total_ratio / dcount,
         stat.zstd_only.total_time_usec / stat.count);
 
     printf(
-        "  [Cloudini-LZ4]  ratio: %.2f time (usec): %ld\n", stat.lossy_lz4.total_ratio / dcount,
+        "  [Cloudini-LZ4]  ratio: %.2f time (usec): %d\n", stat.lossy_lz4.total_ratio / dcount,
         stat.lossy_lz4.total_time_usec / stat.count);
     printf(
-        "  [Cloudini-ZSTD] ratio: %.2f time (usec): %ld\n", stat.lossy_zstd.total_ratio / dcount,
+        "  [Cloudini-ZSTD] ratio: %.2f time (usec): %d\n", stat.lossy_zstd.total_ratio / dcount,
         stat.lossy_zstd.total_time_usec / stat.count);
     std::cout << std::endl;
   }
